refactor(etl): centralize transform contract to etl root
Move contract definitions and build_transform_result function from
transform subdirectory to the etl root directory for better modularity
and easier access by all transform modules. This centralizes the contract
versioning and result building logic.

#130
