--- apps/backend/src/etl/__init__.py
@@ -1,1 +1,0 @@
-from .get_decfec_file import *
